requirements_dev.txt
django-debug-toolbar
django-extensions
django-template-repl
factory_boy
mock
Werkzeug
